CAMEL-13708: Properties component should have API to make it easier to lookup a property on-demand or from all pre-loaded properties.
