+++ core/camel-core/src/test/java/org/apache/camel/component/properties/OptionalPropertiesDslTest.java
@@ -60,7 +60,9 @@ public class OptionalPropertiesDslTest extends ContextTestSupport {
     @Override
     protected CamelContext createCamelContext() throws Exception {
         CamelContext context = super.createCamelContext();
-        context.addComponent("properties", new PropertiesComponent("classpath:org/apache/camel/component/properties/myproperties.properties"));
+        PropertiesComponent pc = new PropertiesComponent();
+        pc.setLocation("classpath:org/apache/camel/component/properties/myproperties.properties");
+        context.addComponent("properties", pc);
         return context;
     }
