+++ core/camel-core/src/test/java/org/apache/camel/component/properties/PropertiesComponentEncodingTest.java
@@ -49,7 +49,8 @@ public class PropertiesComponentEncodingTest extends ContextTestSupport {
     @Override
     protected CamelContext createCamelContext() throws Exception {
         CamelContext context = super.createCamelContext();
-        PropertiesComponent prop = new PropertiesComponent("classpath:org/apache/camel/component/properties/myutf8.properties");
+        PropertiesComponent prop = new PropertiesComponent();
+        prop.setLocation("classpath:org/apache/camel/component/properties/myutf8.properties");
         prop.setEncoding("UTF-8");
         context.addComponent("properties", prop);
         return context;
